backend/requirements.txt
asgiref==3.10.0
awsebcli==3.25.2
bcrypt==5.0.0
blessed==1.22.0
botocore==1.40.58
cement==2.10.14
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
decorator==5.2.1
Deprecated==1.2.18
dj-database-url==3.0.1
Django==5.2.7
djangorestframework==3.16.1
fabric==3.2.2
fonttools==4.60.1
idna==3.11
invoke==2.2.1
jmespath==1.0.1
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
packaging==24.2
pandas==2.3.3
paramiko==4.0.0
pathspec==0.12.1
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pycparser==2.23
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
typing==3.7.4.3
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.14
wrapt==1.17.3
